Fix bug with nested silent comments
Lines weren't being processed inside a nested comment which meant that
the parent node ended up being nil for the child lines. This broke
things when popping back up the stack. Since a push never happens for
the line and therefore the children were empty.

This fix makes comments be processed. That might not be the desired
result since there isn't much point in processing silent comments. But
it is a quick fix :)