Fix for Rails 3.2.3 textarea helpers
Textarea helpers on 3.2.3 emit a newline after the opening textarea tag,
which caused Haml to mess up indenting and add an encoded newline to the
textarea content. This fixes the issue by replacing the first newline
with <haml:newline/> in the Haml buffer, and then changing it to a true
newline at render time.