Require "abstract_controller/rendering" before monkeypatching.
See issues #529 and #519.

If users create a mimimal Rails stack this module may not be have been
loaded.

I'm less than  happy about making Haml require this module, it seems
that if people didn't want to require it in the first place, Haml
shouldn't do it. A better solution than the <haml:newline/> hack would
attack this problem at the root, but for now this should at least solve
the most immediate issue.