Deprecate Ruby 1.8.6 compatibility methods.
`flatten`, `set_hash` and `set_eql?` in `Util` are only used to fix
the corresponding methods in Ruby 1.8.6, which is no longer supported.

Mark these methods as deprecated, and issue a warning if they are used.

Also remove test for these methods - the tests only check that the
core/std-lib Ruby methods are working.

As far as I can tell, these methods are no longer used anywhere in the
Haml codebase - the tests were the only uses.