Provide test details in exception tests
The exception tests in test/engine_test.rb that generate a warning with
-w only show the test filename, with erroneous line numbers, making it
impossible to see what caused the warning.

Provide name of the test in question when rendering for better
identification of what is causing the warnings. Provide only the test
name and not the file name in order to be consistent with other tests.