Check for Rails before using __in_erb_template
The tests all load (some version of) Rails. Haml is not always going to
be used in a Rails environment, and ap_ge3? will be false in these
situations, so __in_erb_template will be used, generating
needless warnings.

Check that Rails is loaded, as well as it being an earlier version,
before using __in_erb_template in generated code.