Conditionally define methods depending on Ruby version
Checking the version over and over again during excecution is wasteful
and pointless; previously there was even a version comparison cache to
speed this up. The proper way to speed this is up is to simply define
the method we need for the given Ruby version when the class is defined.