Make Parser a class and remove from Engine
You can now instantiate your own Haml parser separately from the Haml
engine if for some reason you should want to. Eventually this will be
useful for tools that convert Haml to Erb, for example. However for the
moment the primary motivation is simply to guarantee a stricter separation
of concerns between the parser, compiler and engine.