Use custom Erubis handler subclass for Erb filter
The Rails Erubis handler has added code to reduce the number of method
calls when adding newlines (see
https://github.com/rails/rails/commit/bd308a0d02c4a6459c05770a564405986a663f47).

This uses a new instance variable, which is initialised in the
add_preamble method. Tilt explicitly ignores the preamble from Erubis,
using its own, so this variable is never initialised. This causes
failures with the Erb filter when this variable is referenced.

By using our own custom subclass the variable can be initialised in the
classes initialize method, preventing failures.

Fixes #669