Fix interpolated comments breaking tabulation
When interpolating single line comments, use the nuke_inner_whitespace
option to prevent format_script from decrementing the @real_tabs
variable in the buffer, which can result in issues later in the Haml.

This is admittedly a bit of a hack, cased by the number of params for
format_script and the number of routes through that method. Other parts
of the compiler explicitly emit code to add the extra tab to take this
into account, which is arguably worse.