Reproduce original behavior of true id/class
Haml 4 renders it as id='true' and class='true' because it's implemented like this:
https://github.com/haml/haml/blob/4.0.6/lib/haml/compiler.rb#L431-L436

For safe Haml 5 release, let's keep the original behavior.
But I think this is not an expected behavior. Thus not adding tests.

Feel free to revert this change after Haml 5 release.
But it may be faster to keep this change and no one depends on this
behavior...

The next fix for this problem should be having a boolean attributes list.