Don't push temple into `@to_merge`
The former implementataion may potentially hit a bug because `@to_merge`
is also modified by `rstrip_buffer!` and it may be hard to implement
`rstrip_buffer!` for all temple expressions. With this commit, the bug
can be avoided.

By the way, for `rstrip_buffer!`, we should maintain `@to_merge`
expressions like `[:text, ...]` and `[:script, ...]` as intermediate
representation between Haml AST and Temple expressions.

Of course I want to drop `@to_merge`, migrate [:text] to [:static] and
carve out `rstrip_buffer!`'s work to a Temple filter.
But it's hard to keep current behavior with that change, mostly for
`_hamlout.rstrip!`. It's dynamic rendering like pretty mode and hard to
reproduce the behavior statically.

I think the behavior requiring `Haml::Buffer#rstrip!` should be removed
in the future. For Haml 5, let's keep the current behavior.