Docs: Help people upgrade to 5 by classifying changes
By identifying breaking changes in the changelog, it will be much
easier for people to upgrade to 5.

Not being a HAML expert, I may have missed some breaking changes,
but this should be a good start.

The distinction between added/fixed is much less important in a
major version bump.

[ci skip]