Fix Erubi on Rails < 5.1
ActionView does not support Erubi just because it is included in the bundle. This checks for Rails to support Erubi rather than only looking to see if Erubi is in the bundle at all.

This change is necessary because `Haml::ErubiTemplateHandler` inherits from `ActionView::Template::Handlers::ERB::Erubi`, which is not actually a useable class in Rail prior to 5.1.