Do not require rails when haml is required
Still load haml/railtie if Rails is already loaded. This returns
to the haml 5 behavior.

Outside of a Rails application, loading Rails is probably not
desirable. It requires significant additional memory and makes
incompatible changes to existing classes.

Inside of a Rails application, Rails is already loaded, so this
change should be a no-op.

This can change behavior for applications that require haml and
then require rails. Previously, that would load the haml railtie,
but with this change, it does not. Effected applications should
be updated to explicitly require haml/railtie after requiring
rails, or require rails before haml.

Fixes #1199